SurfProfilePageTests and CustomerMethod

+++ src/test/java/RegistrationPageTests.java
@@ -16,51 +16,50 @@ public class RegistrationPageTests {
 
     private WebDriver driver;
     private RegistrationPage pageObject;
-    private CustomerClient customerClient = new CustomerClient();
     private String expected;
     private String token;
     private String email;
     private String password;
-    private String name;
+    private String errorMassage = "Клиент не зарагистрирован, войти в аккаунт не удалось, accessToken равен null ";
     @Before
     public void startUp() {
         driver = new ChromeDriver();
         pageObject = new RegistrationPage();
         driver.get(Links.getRegistrationWindow());
         email = CustomerGenerator.randomEmail();
-        name = CustomerGenerator.randomName();
         password = CustomerGenerator.randomPassword(8);
-        pageObject.setNameInRegistrationForm(name);
-        pageObject.setPasswordInRegistrationForm(password);
-        pageObject.setEmailInRegistrationForm(email);
+        pageObject.setNameInRegistrationForm(driver, CustomerGenerator.randomName());
+        pageObject.setPasswordInRegistrationForm(driver, password);
+        pageObject.setEmailInRegistrationForm(driver, email);
     }
     @Test
     public void checkRegistration() {
-        pageObject.ButtonRegistrationClick();
+        pageObject.ButtonRegistrationClick(driver);
         WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(1));
         wait.until((WebDriver d) -> d.getCurrentUrl().equals(Links.getPersonalAccountWindow()));
         expected = driver.getCurrentUrl().toString();
         Assert.assertEquals("Не открылась страница - " + Links.getPersonalAccountWindow(), expected, Links.getPersonalAccountWindow());
 
         token = CustomerClient.loginCustomer(email, password);
-        Assert.assertNotNull("Регистрация не прошла, accessToken равен null", token);
-        CustomerClient.deleteCustomer(token); ;
+        Assert.assertNotNull(errorMassage, token);
     }
     @Test
     public void checkRegistrationShortLogin() {
         password = password.substring(0, 6);
-        pageObject.ButtonRegistrationClick();
+        pageObject.ButtonRegistrationClick(driver);
         WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(1));
         wait.until((WebDriver d) -> d.getCurrentUrl().equals(Links.getPersonalAccountWindow()));
         expected = driver.getCurrentUrl().toString();
         Assert.assertEquals("Не открылась страница - " + Links.getPersonalAccountWindow(), expected, Links.getPersonalAccountWindow());
 
         token = CustomerClient.loginCustomer(email, password);
-        Assert.assertNotNull("Регистрация не прошла, accessToken равен null", token);
-        CustomerClient.deleteCustomer(token);
+        Assert.assertNotNull(errorMassage, token);
     }
     @After
     public void teardown() {
+        if (token != null){
+            CustomerClient.deleteCustomer(token);
+        }
         this.driver.quit();
     }
 }
